Moved new Blueprints creation to ContentAdder logic
