+++ TabletopTweaks/TabletopTweaks.csproj
@@ -280,6 +280,7 @@
     <Compile Include="NewContent\Archetypes\NatureFang.cs" />
     <Compile Include="NewContent\Archetypes\DivineCommander.cs" />
     <Compile Include="NewContent\Archetypes\ElementalMaster.cs" />
+    <Compile Include="NewContent\Bloodlines\BloodragerArcaneBloodline.cs" />
     <Compile Include="NewContent\Classes\Cavalier.cs" />
     <Compile Include="NewContent\Classes\Loremaster.cs" />
     <Compile Include="NewContent\Archetypes\MetamagicRager.cs" />
